introduce `mockDesignSystem`
This only exists to implement a poor man's dependency injection system.
We essentially want to reuse the existing `designSystem` as much as
possible. But for some migrations we want to do things slightly
differently (e.g.: not mutating arbitrary values)

This function allows us to override some of the `designSystem` methods.

Bonus points: overrode the `designSystem#parseCandidate` such that it
doesn't do any caching anymore. This means that we can safely use
`designSystem.parseCandidate` instead of using
`parseCandidate(candidate, designSystem)`