hide skipped tests
When you have a lot of tests, and use a `.only` on a specific test, then
all other tests are still visible in the UI. If you are logging
something, then you always have to scroll up.

To prevent this, we can use this flag to always hide the list of
skipped tests.

Note: the UI itself still has an indication that _something_ is being
skipped:
```console
 Test Files  1 passed (1)
      Tests  1 passed | 97 skipped (98)
   Start at  13:27:14
   Duration  55ms
```