inject the border compatibility CSS in "root" files
We were always injecting the border compatibility CSS above the first
`@layer base`. But in projects where you have multiple files using this
`@layer base` it means that the reset is injected in each file.

Instead, only try to inject it in files where `@import 'tailwindcss'` or
`@import 'tailwindcss/preflight'` is being used.

When we see a file with `@import 'tailwindcss'`, we consider it a root
file because that's where Tailwind CSS will be injected.

If you split up your code, it could be that you have a different file
that contains `@import "tailwindcss/preflight";`. This is the import
responsible for adding the `@layer base` resets, so this is a good file
to put the border compatibility CSS as well.