merge `sources` and `detectSources`
We will be applying auto source detection on normal sources found via
`@source` as well. This means that we don't need a distinction between
sources and detect sources.

The only difference would be that in the clients (`@tailwindcss/vite`,
`@tailwindcss/postcss` and `@tailwindcss/cli`), you pass everything as a
single array of sources.