add package manager abstractions
To make sure we are installing in the correct location (`dependencies`
vs `devDependencies`) we need to provide some flags to the package
manager.

In a perfect world the flags used by npm, pnpm, bun and yarn would be
the same but that's not the case. This abstraction allows us to use a
consistent interface where we can add and remove files and point the
location we want.