use existing `closingBracketStack`
We still apply the same rules where we don't really care about special
characters such as `{`, `}` or `;` when we are inside of `(` and `)`.

We can safely look at the last character in the `closingBracketStack`
because when we are in parentheses, we don't update the
`closingBracketStack` until we see a closing `)`.