add `hoistStaticGlobParts`
This is an implementation similar to what we already have in Rust. The
idea is that we want to move the static parts to the `base` path. This
allows us to use `..` in the pattern and the glob would be adjusted
correctly.

Without this, globby will error if your pattern escapes the base path by
using `..` in the pattern.